Moves viewer components to `components/documents`
Moves Paginator from `@/common` to `lib/components/common`

Fix paginator import

Fix paginator
