+++ src/lib/components/documents/sidebar/Actions.svelte
