+++ src/lib/components/documents/sidebar/Data.svelte
