+++ src/lib/components/documents/sidebar/DocumentMetadata.svelte
