+++ src/lib/components/documents/sidebar/Notes.svelte
