+++ src/lib/components/documents/sidebar/Sections.svelte
